Create products.jsp
Look into lines 68-77, the cforEach loops wont work unless some stuff is imported.
I was trying to make a script that writes more html code, but have no clue how to get it working. Tried so many things lol. But the goal here is to make the program write in the code dynamically, so when we sort and stuff we can remake the tables no problem.